refactor to use usize instead of u64

+++ src/constants.rs
@@ -46,15 +46,15 @@ pub mod cs2 {
 }
 
 pub mod elf {
-    pub const PROGRAM_HEADER_OFFSET: u64 = 0x20;
-    pub const PROGRAM_HEADER_ENTRY_SIZE: u64 = 0x36;
-    pub const PROGRAM_HEADER_NUM_ENTRIES: u64 = 0x38;
+    pub const PROGRAM_HEADER_OFFSET: usize = 0x20;
+    pub const PROGRAM_HEADER_ENTRY_SIZE: usize = 0x36;
+    pub const PROGRAM_HEADER_NUM_ENTRIES: usize = 0x38;
 
-    pub const SECTION_HEADER_OFFSET: u64 = 0x28;
-    pub const SECTION_HEADER_ENTRY_SIZE: u64 = 0x3A;
-    pub const SECTION_HEADER_NUM_ENTRIES: u64 = 0x3C;
+    pub const SECTION_HEADER_OFFSET: usize = 0x28;
+    pub const SECTION_HEADER_ENTRY_SIZE: usize = 0x3A;
+    pub const SECTION_HEADER_NUM_ENTRIES: usize = 0x3C;
 
-    pub const DYNAMIC_SECTION_PHT_TYPE: u64 = 0x02;
+    pub const DYNAMIC_SECTION_PHT_TYPE: usize = 0x02;
 }
 
 pub const GRENADE_FILE_NAME: &str = "grenades.json";
