+++ src/cs2/entity/inferno.rs
@@ -7,11 +7,11 @@ use crate::cs2::{
 
 #[derive(Debug, Clone, PartialEq)]
 pub struct Inferno {
-    controller: u64,
+    controller: usize,
 }
 
 impl Inferno {
-    pub fn new(controller: u64) -> Self {
+    pub fn new(controller: usize) -> Self {
         Self { controller }
     }
