+++ src/cs2/entity/molotov.rs
@@ -7,11 +7,11 @@ use crate::cs2::{
 
 #[derive(Debug, Clone, PartialEq, Serialize)]
 pub struct Molotov {
-    controller: u64,
+    controller: usize,
 }
 
 impl Molotov {
-    pub fn new(controller: u64) -> Self {
+    pub fn new(controller: usize) -> Self {
         Self { controller }
     }
